feat: Add additional functionality to element.sh
This commit expands the behavior of the element.sh script. Now, instead of simply
displaying an error message when an element is not found in the database, the script
reads the element data into variables and outputs a comprehensive description of the
element. This includes the atomic number, name, symbol, mass, melting point, boiling
point, and type.